requirements.txt
opencv-python
anthropic
flask
requests
pillow
numpy
